Code-dedupe `sub`, `gsub`, and `ssub` verbs
